LayoutTests/imported/w3c/web-platform-tests/workers/SharedWorker_dataUrl-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'worker.port.onmessage = e => {
  parent.postMessage(e.data, '*');
}')
CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'worker.port.onmessage = e => {
  parent.postMessage(e.data, '*');
}')


Harness Error (TIMEOUT), message = null

TIMEOUT Data URL not shared by cross-origin SharedWorkers Test timed out
NOTRUN Data URLs shared by same-origin SharedWorkers
